* Add a "release with debug" profile.
Running any meaningful instance requires the solver to be compiled
in release mode for performance. A downside is that now debug symbols
are stripped, and any bugs are more difficult to track down.

This commit adds a new profile, which compiles the solver in release
mode, but includes debug symbols. Even though some symbols might be
useless due to optimizations the compiler applied, it does help in
certain situations. This does not impact performance of the solver,
only binary size.

Approved-by: Emir Demirović