Merged in feature/cardinality-network-encoding (pull request #5). But there is an issue: the cardinality network encoding does not do preprocessing which can cause crashes. This issue will be fixed later with a more detailed refactoring of encoders.
Cardinality network encoding

* Handle empty functions

* Fix tests after rebase

* Renamed `create` to `from_function`, and added documentation to the
public methods of the cardinality network encoder.

* Rename UpperBoundEncoding::CN -> UpperBoundEncoding::CNE to be
consistent with the GTE option.

* Make the encoder use the correct method to retrieve the function as a
collection of weighted literals.

* Added similar logging to CNE as is present in GTE

* Fix build error after rebase.

* Make logging consistent with GTE

* Run cargo fmt

* Encapsulate the upper bound encoder inside LinearSearch.

If in future the search procedure changes to e.g. a core-guided
method, then having the upper bound encoding as a field on the
Pumpkin struct makes little sense. Even though this approach
allocates the encoding on the heap, this allocation does not happen
on the hot-path, so it should not be an issue.


Approved-by: Emir Demirović