Merged in feature/maxsat-tests (pull request #31)
Add tests for MaxSAT instances

* Add more wcnf instances

* Remove testing temporary files from repository

* Ignore testing artefacts

* Make the test files referenceable from rust identifiers

* Remove unused import

* Merge branch 'master' into feature/maxsat-tests

* Make changes to build on Windows.

The references to popen and pclose were removed, since we do not use the
pipe functionality of the checker. This also means the sys_utils.* files
can be removed.

* Remove WCNF instance with non-unit soft clauses

* Revert "Remove WCNF instance with non-unit soft clauses"

This reverts commit 019ae0c5a0f1c28cf2f76631c068afd8f380223b.

In hindsight, we should include such test cases since it is important to
support the entire benchmark suite.

* Fail with non-zero exit code if the objective value is incorrect.


Approved-by: Emir Demirović