Merged in refactor/event_sink (pull request #42)
Improve EventSink implementation

* Cache the present events in the event sink for easier reading.

Profiling showed >80% of run time was spent in EventSinkIter::next().
This is because the iterator had to always go through all the domains to
ensure no events occurred for each of them.

Now, we maintain a vector that has all the events and can be easily
drained, and only use the bit vector to indicate whether an event was
already triggered in the sink.

* Add enumset dep, and use in event sink


Approved-by: Jeff Smits