This PR starts the work of the FlatZinc frontend for Pumpkin. Functionally, we are able to run the nqueens model from the MiniZinc benchmarks and produce an output that conforms to the FlatZinc specification.
Since implementing the entire FlatZinc specification is a lot of work, doing so incrementally through multiple PR’s is probably easier to manage with reviewing the code. Hence, this PR only implements support for the int_lin_ne constraint, and does not handle optimisation models either. In any case, the solver should crash if it encounters FlatZinc it does not understand. This is to make explicit that support is yet to be added.

Merged in feature/flatzinc-v2 (pull request #40)
Approved-by: Jeff Smits