Merged in feature/commit-hooks (pull request #66)
Share a pre-commit hook that runs cargo fmt and cargo clippy before committing, to catch code that will be rejected by the CI pipeline.

Because git hooks are local to every user, they are shared in a separate .githooks folder. The scripts in that folder can be symlinked to the .git/hooks directory to be “installed”. A project-level README is added that details the command to create the symlinks.

Approved-by: Jeff Smits