Merged in feature/negative-domain-bounds (pull request #69)
Implement support for negative domain bounds in the mediator.

Even though we could implement domains on the assumption that the lower
bound is non-negative, this makes writing the FlatZinc parser a lot more
complicated as everything would have to be converted to views, to offset
possible negative lower bounds. As the signature of create_domain
already takes signed integers, it implies that negative numbers are
possible anyways. Hence, we support it.

Approved-by: Emir Demirović