Introduce an abstraction for constraints.
They are implemented using propagators, possibly as a decomposition. The constraints are clearly documented, and can be reused in the library examples, as well as models implemented using pumpkin as a library.

Merged in feature/constraints (pull request #67)

PUM-67 Add constraint abstraction

* Extract int_lin_le_reif constraint and refactor the
disjunctive_scheduling example to use it.

* Merge branch 'master' into feature/constraints

* Update CHANGELOG

* Merged master into feature/constraints
* Make constraints API use `impl Into<>` where possible, to make the API
more flexible.

* Merge branch 'feature/constraints' of bitbucket.org:EmirD/pumpkin-private into feature/constraints

* Add type annotations to fix build

* Refactor constraints to extension trait pattern.

* Merge remote-tracking branch 'origin/master' into feature/constraints

* Merged master into feature/constraints

Approved-by: Jeff Smits