Merged in feature/extend-collected-data (pull request #63)
Add the additional collection of the following statistics:
- Average conflict size
- Average learned clause length
- Time spent in the actual solver in ms
- Average backtrack amount

Logs the statistics using the info! logging in the format x {statistic name}: {statistic value} where {statistic value} is assumed to be a float or an integer at the moment and {statistic name} is assumed to consist solely of whitespaces and characters. It only performs this logging when solving via the OptimisationSolver and LinearSearch. 

info! no longer provides timestamp/call-site.