Merged in bugfix/mzn-redefinition (pull request #109)
Remove redefinition of set_in_reif constraint.

* Remove redefinition of set_in_reif constraint.

Previously there was a redefinition which explicitly aborted when the
`set_in_reif` constraint was used and the set was a variable. However,
because we include `nosets.mzn`, this situation will never occur. In
fact, we redefined the predicate when it was already redefined in
`nosets.mzn`, so minizinc complained.


Approved-by: Imko Marijnissen