Merged in refactor/propagator_docs_and_usage (pull request #145)
The idea is to streamline the process of propagator development, especially for newcommers. Changes:

Added more documentation on how to implement a propagator, that serves as a nice checklist/guideline for propagator development. See the mod file in propagation.

Added default implementations to the Propagator trait to make it easier to get the first working propagator version, and some supporting documentation.

Moved propagator documentation from the propagator folder to the propagation folder. The propagator folder now holds only concrete implementations, whereas propagation explains the entire concept and provides guidelines on how to implement propagators.

A side effect is that debug_propagate_from_scratch now requires providing explanations for its propagations, because it is used in the default implementation of propagate.

Approved-by: Maarten Flippo