Merged in feature/alternating-brancher (pull request #179)
Switches between two branchers based on certain criteria.

E.g. switches every solution or switches after the first solution. This behaviour can be useful in the context of scheduling where we would like to find a “good” first solution and then let VSIDS take over.

At the moment, the brancher switches between VSIDS and the provided brancher, this can be changed to switching between 2 provided branchers in the future if that is more appropriate.

Approved-by: Maarten Flippo