Remove PropagatorVariable as a concept.
Previously, the propagator variable was a combination of LocalId and a
variable. We paired these up for use in propagators because the old
design for lazy explanations queried the propagator for an explanation
of a propagation performed on a LocalId, which meant every propagation
was connected to a LocalId.

Since the new design, where lazy explanations are closures created
when the propagation is performed, this design is no-longer necessary.
We saw this because `PropagatorVariable<Var>` was simply a wrapper
around `Var`. To reduce complexity and lines of code, we remove this
artefact of days gone by.