Merged in feature/reified-constraints (pull request #192)
This PR adds the ability to half-reify any propagator. It introduces the ReifiedPropagator, which can wrap any other propagator to reify.

To support this, some changes were made to the Propagator trait. Most notably, initialise_at_root now no longer propagates. It is only there to set up the internal state of the propagator. The solver will call Propagator::propagate after Propagator::initialise_at_root when a new propagator is added to the solver.