Merged in refactor/integration-tests (pull request #194)
Previously, the integration tests would, inside the test, build the pumpkin-cli binary in release mode. This was done to improve the speed of the solver under test. However, there is a big problem with this approach: Since multiple tests are run in parallel, and each test would run cargo build --release, the individual tests were competing for the compilation process to finish. This causes massive slowdowns.

Luckily, when running cargo test on a binary crate, the binary is built as well. Further, we can get the path to the binary through environment variables set by cargo during compilation. With that in mind, we can simplify the test setup so that now the tests do not individually try to build the solver executable. Instead, we can assume it is already built.

Approved-by: Imko Marijnissen