Merged in bugfix/restore_at_root_satisfy (pull request #198)
Verify optimality WCNF + Fixing Restoration at root

The aim of this PR is three-fold:

We identified that the solutions we are reporting are not optimal. Luckily, the maxsat checker we use to check the solution and solution cost allows us to specify an expected optimal value. If the values don’t match, the test fails. This means for every wcnf instance we need to specify the optimal objective value, which we can compute with other solvers.

We also noticed that the on_encoding_objective_function was incorrect and unclear; now we assume that the variable/value selector has support for the addition of new variables rather than having a separate method in the interface.

The Solver was not restoring to the root when a solution was found and it was not notifying the brancher when a solution was found when solving under assumptions


Approved-by: Maarten Flippo