Add fix for array of bools being parsed as array of int (#9)
This fixes the issue where the following fails to parse:

```
var bool: x1 :: output_var;
var bool: x2 :: output_var;
var bool: x3 :: output_var;

solve :: bool_search([x1, x2, x3], input_order, indomain_max) satisfy;
```

This was due to the fact that the flatzinc crate which we use does not know what the type of the provided array is and it then automatically parses it to an array of ints. The fix is to parse int arrays when parsing boolean array but only allowing boolean values in there.