Notification upon backtracking (#10)
This PR aims to add notification upon backtracking which could lead to
more efficient propagators. Currently, whenever backtracking occurs, the
propagator does not know which changes have been undone; this could lead
to significant effort in recalculating from scratch even when a small
adjustment to the data structures could have been necessary.

The PR aims to add this by retaining the same structure as the normal
event sink but with backtrack events (this could likely use the same
events but if we want to provide more information in the future then it
will likely be useful to have separate events).

---------

Co-authored-by: Maarten Flippo <maartenflippo@outlook.com>