(Small PR) Fixing boolean flags CLI + adding statistic printing to (Max)SAT (#16)
Setting the default value of a flag to true resulted in the flag always
being true, regardless of whether the flag was present or not; this has
been addressed by adding the prefix `no-` to the learning minimise flag
and negating its value.

Additionally, statistics were not even printed with the flag turned on
for (Max)SAT problems, this has been added.