Incremental linear not equal (#14)
This PR aims to add incrementality to the linear not equal propagator.
The idea is to make it as simple as possible while retaining the core
functionality.

The propagator now keeps track of how many variables are fixed and the
value of the left-hand side; the current implementation can only
propagate if there is a single unfixed variable. A conflict can only be
detected if all of the variables are fixed.