Remove PropagatorConstructor (#33)
Closes #32

With the removal of `PropagatorVariable`, the presence of
`PropagatorConstructor` is redundant and serves only to confuse people
when implementing a new propagator.

This PR aims to address this issue by removing the
`PropagatorConstructor` in its entirety and moving the registration of
variables to the `initialise_at_root` method using the new
`PropagatorInitialisationContext`. This simplifies the implementation of
a new propagator significantly.