Extend cumulative explanations (#6)
As stated in ["Improving scheduling by learning - Schutt
(2011)](https://minerva-access.unimelb.edu.au/items/fe4825ac-cd4a-582c-bacc-d4ae384f3f76);
there are 3 types of explanations for the cumulative constraint. The
explanations which we currently use for propagation explanations are
non-standard; the predicates concerning the propagating task are lifted
but the rest is not while the conflict explanation is simply the naive
explanation.

This PR aims to address the aforementioned issue by adding the naive,
big-step and pointwise explanation to the cumulative and adding an
option which allows the user to select which one to use via a
command-line option. Note that this PR does not generate a chain of
propagations as is done in the PhD thesis but it does contain the
prevention of recalculation of explanations of the same profile by
caching.

---------

Co-authored-by: Imko Marijnissen <icwmmarijnisse@tudelft.nl>
Co-authored-by: Maarten Flippo <maartenflippo@outlook.com>