Fix issue with linear not equals not being enqueued + Wrong Lazy Reason (#24)
There were 2 mistakes in the not equal propagator:
- The notify method checks whether the context is fixed before updating
the number of fixed variables; this could be wrong due to the
notification taking place after backtracking. Removing this check fixes
this issue
- The `unfixed_variable_has_been_updated` was wrongly set to true if no
update took place; it could be argued that the variable should be true
but then we do not currently detect the case when this value becomes
part of the domain again (which could be due to a bound update). This
issue was fixed by only setting this variable if the propagator removes
the variable.

Additionally, oftentimes the propagators assume that they first get
notified of all the "forward" events (i.e. tightening of the
bounds/removals/assignments) and then of the backtrack events. This
issue is now fixed by first processing the "forward" events when
notifying the propagators of the backtrack events.

**Update**
This PR also addresses the issue mentioned in #25 by eagerly calculating
the reason rather than relying on the context during lazy reason
computation.