Removing restarts when search is static (#30)
Currently, the search is restarted even if the search strategy is
static; this would mean that, in certain cases, a lot of work is being
redone.

This PR aims to address this by adding a new function for `Brancher`s,
`VariableSelector`s, and `ValueSelector`s named `is_restart_pointless` which
returns a boolean value indicating whether it changes during the search
or not (with a default implementation returning `true`). This allows us
to only restart if the current brancher is not static; i.e. if
performing a restart _could_ lead to different decisions.

Additionally, this PR adds a flag for turning off restarts altogether.