Start of the python interface (#35)
This PR starts the introduction of the python interface. Since it is a
big addition, I broke it up into smaller amounts for review. This PR
sets up the project and implements the bare minimum to run an nqueens
model from python (see the included sample).

There are still some remnants from the project scaffold. In particular
the `tests` directory in `pumpkin-py`. However, I do not yet know how to
test the python interface beyond duplicating many of the tests we
already have. I have started an issue #34 to discuss testing this code,
so we don't have to focus on that in this PR.