Re-introducing checking of Reason (#49)
For a while, the code for checking the reason for a propagation has been
unused; likely due to the fact that the reasons which were provided to
the method were not `PropositionalConjunction`s anymore.

This PR aims to address this by calculating the reason eagerly (if the
assertion levels are set to extreme) and checking these reasons. This
already led to an error becoming apparent in the integer multiplication
propagator due to an off-by-one error!