Add option to select cumulative selection method (#45)
Currently, it is only possible to create the cumulative using
`TimeTableOverIntervalIncremental` and only that type of reasoning.

This PR aims to address this by allowing the user to pass a flag which
selects the approach which is used to propagate the cumulative
constraint (currently constricted to time-table reasoning but it might
be extended in the future!).

_Additionally, the integration tests can now be easily extended to
include the different types of reasoning (and the different types of
explanations), this is also addressed by this PR_

---------

Co-authored-by: Imko Marijnissen <icwmmarijnisse@tudelft.nl>
Co-authored-by: Maarten Flippo <maartenflippo@outlook.com>