Getting predicate literals in python (#59)
This PR implements the equivalent of the `solver.get_literal` function
in the Python interface. We also implement casting boolean expressions
to integer ones. Variable names are forwarded as appropriate, although
associating names with predicate literals is not supported yet.

Even though we will at some point move away from explicit literals, the
API we have here (or an evolved version of it) will still show up in the
revamped solver version.