Adding the logging of statistics for propagators (#56)
Currently, we only print statistics concerning the solver itself,
however, it would be useful if we at least had the option to print
statistics for propagators.

This PR addresses the aforementioned issue by adding a new method on the
propagator struct which can print any relevant statistics collected by
the propagator. _It also refactors some of the structures to be nicely
formatted in the correct place_

---------

Co-authored-by: Maarten Flippo <maartenflippo@outlook.com>