fix: float stat into one line (#83)
When testing statistics, I found float statistics to be divided into
multiple lines:

 _engine_statistics_time_spent_in_solver=139
 _learned_clause_statistics_average_conflict_size=2
 _learned_clause_statistics_average_conflict_size=.
 _learned_clause_statistics_average_conflict_size=247863247863248

_learned_clause_statistics_average_number_of_removed_literals_recursive=2

_learned_clause_statistics_average_number_of_removed_literals_recursive=.

_learned_clause_statistics_average_number_of_removed_literals_recursive=5555555555555554
The origin lies in the writer for a float: it writes it to the writer in
multiple parts (core::fmt::write_formatted_parts). Currently, each of
these parts would be interpreted as being a statistic, resulting in
multiple lines.

I now removed the Write trait from StatisticLogger alltogether, and
simply called the required functions directly. Given that log_statistic
directly writes to stdout anyway, I don't think this compromises the
extendability of the logger.