fix: complete proof when propagator initialization identifies root-level conflict (#80)
Closes #79

Currently, when a root-level conflict is found when adding a propagator,
the state of the propagator does not contain the reported conflict.

However, this means that the `complete_proof` step cannot properly
create the final learned clause; this PR addresses this by properly
setting the state of the `ConstraintSatisfactionSolver` to be
conflicting using the conflict returned by the propagator.