feat: incremental backtracking for the cumulative constraint (#60)
Currently, the cumulative constraint simply recalculates all structures
from scratch upon backtracking, even if only small updates occurred.

This PR implements incremental backtracking by making use of the
backtracking notification system. Whenever a propagation occurs, it is
determined what (parts of) the mandatory part should be removed and
updates those parts only. This allows for a constant number of
adjustments performed for each task upon backtracking.

A flag can be passed to the solver which specifies whether to use this
incremental backtracking or not (default is that it doesn't use the
incremental backtracking).

_It also splits into different structs the parameters passed to the
propagator from the dynamic structures which are adjusted through search
(e.g. the stored bounds or the stored updates for the tasks)_

---------

Co-authored-by: Imko Marijnissen <icwmmarijnisse@tudelft.nl>
Co-authored-by: Maarten Flippo <maartenflippo@outlook.com>