feat: adding synchronisation for the cumulative constraint (#86)
It is difficult to compare the different types of changes made to the
cumulative due to the fact that traces oftentimes diverge due to
differences in the explanation(s)

This PR addresses this issue by creating a synchronisation between the
incremental and the non-incremental version of the time-tabling
propagators by ensuring that the explanations for the conflicts and the
propagations are the same as would be found by the non-incremental
propagator and the incremental version of the propagator.

For both the per-point reasoning and the over-interval reasoning, a
`synchronisation` module is created which finds the correct conflicts
and orders the profile tasks in the profile in the correct order!

_Edit: After discussion with @maartenflippo, this PR now also contains a
new crate which contains procedural macros; these are currently used to
create test cases for the cumulative where we want to test a lot of
combinations of tests!_

---------

Co-authored-by: Imko Marijnissen <icwmmarijnisse@tudelft.nl>
Co-authored-by: Maarten Flippo <maartenflippo@outlook.com>