fix: wrong profile found for over-interval conflict explanation (#111)
Currently, the synchronised conflict explanation simply uses the first
conflicting profile as its explanation. While this is correct in the
sense that the start point of this profile is the same as the start
point of the profile which would have been found by the non-incremental
propagator, it suffers from the issue that the first conflicting profile
does not necessarily span all of the adjacent profiles which have not
been merged.

This PR aims to address this issue by finding the first profile which is
conflicting **and** then creating a profile with all subsequent profiles
which could be merged.