feature: lazy explanation context (#123)
When using lazy explanations, the propagator receives the internal
`Assignments` structure. However, we want them to access more than just
the assignments. So we introduce a dedicated context to expose a wider
API.

Since it is likely undesirable that propagators can access the
`Assignments` directly, the accessor is marked deprecated. The only spot
where we use this is now explicitly allowed but may need to be
refactored.