fix: post process new domain (#129)
When creating a new domain, it could be that the lb/ub is a hole, or
that there are holes outside of the bounds. This is currently not post
processed before creating the domain, so this happens after minimisation
for the first time. This can cause the number of literals after
minimisation to be higher than before, causing an overflow:

https://github.com/ConSol-Lab/Pumpkin/blob/d66ecb6e541b9d15562dc7db223b8a86b3bade13/pumpkin-solver/src/engine/conflict_analysis/minimisers/recursive_minimiser.rs#L65

Post-processing the new domain fixes this problem.