feat: no additional variables during constraint construction (#131)
There are three spots in the flatzinc parser & constraint constructors
that I believe to contain unnecessary variable creation:
- `create_domains` for `BooleanLessThanOrEqual`  and `BooleanEqual`
- `compile_bool2int`

It seems like we can directly used the parsed literals instead of
creating a new variable and a linking clause. It only requires access to
the underlying `integer_variable` of the literal, but I don't see that
as a problem. Is there any reason for doing it this way, or is it
because of the legacy structure?