feat: improve element propagator (#140)
Re-opening https://github.com/ConSol-Lab/Pumpkin/pull/130 due to a
change in my repository.

The `element` propagator improves the bounds of the RHS based on the
items in the array. For this, it takes the full array into account.
However, it could be that some of these items are not possible anymore
due to a (partial) assignment of the `index`. The propagator doesn't
currently take this into account, and I believe this can be improved.