fix: disjunctive_scheduling example had incorrect negation constraint (#114)
Hi, another tiny fix while getting everything running. 

I'm pretty sure the negation of `literal => s_y - s_x <= -p_y` is as
follows:

$$l => s_y - s_x <= -p_y$$
$$\neg l => s_y - s_x > -p_y$$
$$\neg l => s_y - s_x >= -p_y + 1$$
$$\neg l => -s_y + s_x <= p_y - 1$$

Also the literal was not actually negated in the second case, so I added
that as well in the way I think you're supposed to do it.