feat: stateful integer (#125)
A well-known concept in CP is variables which are automatically updated
to their previous values by the solver.

This PR (based on [this
PR](https://bitbucket.org/EmirD/pumpkin-private/pull-requests/58)) aims
to implement these variables by creating a new trail for changes to
these stateful variables. Upon backtracking, these changes are undone.

As a potential application area, I have implemented it in the
`LinearLeq` propagator such that it does not have to recalculate the
incremental state from scratch.