feat: add method for retrieving relevant brancher events (#138)
For some instances (e.g. `2023/code-generator/unison.mzn`, with
`2023/code-generator/mips_gcc.flow.find_regno_partial.dzn`), it showed
up in profiling that the dynamic brancher went through a lot of
branchers upon certain events (in this case the `on_unassign_integer`
event, even though only a single brancher was interested in these
events).

To alleviate this issue, this PR adds a method for `Brancher`s,
`VariableSelector`s, and `ValueSelector`s which allows them to indicate
which events they are interested in; this allows the `DynamicBrancher`
to only call the `Brancher`s which are interested in it.