fix: allow reification of lazy reasons (#152)
Previously, submitting a lazy reason in a reified propagator would crash
the solver. Implementing this is non-trivial, as reasons used in
conflict analysis were returned as slices. However, we need to insert a
reification literal into that slice somehow.

One approach would be to always return some owned buffer like a `Vec`
from the reason store, so that it can insert the reification literal
there. However, that would entail allocating when we ideally want to
avoid that.

Instead, the solution implemented here is to provide the buffer that the
reason should be written in to. Then it is up to the caller to either
allocate or reuse an existing allocation.