fix: explain root-level assignments in the proof log during conflict analysis (#163)
When pumpkin is logging a full proof during solve, we were ignoring the
root-level assignments that are encountered during conflict analysis.
Just as in other areas where we do root-level explanations, we can
assume the root-level assignment either has a 'unit nogood step id' or
it is implied by a sequence of 'unit nogood step ids'.