refactor: Renamed "stateful" into "trailed" (#178)
I want to document TrailedIntegers, but I noticed a few inconsistencies
with the name. Before documenting, I decided to first do a rename, so
that it is easier to review the PR. This PR only contains name changes.

We had the struct called TrailedInt, but in the code, variables were
named "stateful_integers". I remember seeing this as a PR comment to
change "stateful" into "trailed", and this was done for the struct name,
but not the variable names. So this seems to be a leftover.

The changes here are:

- StatefulAssignments -> TrailedValues
- StatefulPropagationContext -> PropagationContextWithTrailedValues
(lengthy name, but I think it is clearer than calling it stateful)
- stateful_assignments -> trailed_values
- TrailedInt -> TrailedInteger
- I added TrailedInteger is in a separate file. Another option is to
have TrailedInteger and TrailedChange all in the same file
(TrailedValues). Previously, we had TrailedChange separate, but
TrailedInteger in the TrailedValues, so at least this way it is a
bit more consistent.
---------

Co-authored-by: Emir Demirović <edemirovic@tudelft.nl>