refactor: Remove nogood module + moving LBD computer to engine module (#185)
The struct for computing LBDs was the only file in the folder "nogoods".
I propose to delete the folder "nogoods" and move the LBD in its parent
folder (engine).

We might also want to rename the struct, since "Lbd" sounds strange.
"LbdComputer" would be more natural. But then "compute_lbd" could be
renamed to "compute". But all of this looks a bit excessive, since in
the code we would have "lbd_compuer.compute(nogood)". For the time being
I decided not to touch the naming, we could revisit this in the future,
unless you have a strong preference to do it now.

On this note, I suppose eventually we would move the Lbd computer to
some other folder. We would probably want to combine "restart strategy"
with our branchers. To be discussed!