refactor: Moved variable names into the engine module (#186)
We had variable_names.rs be in the top folder, right next to lib.rs. It
seemed appropriate to move it into the engine folder.

We could document it in the future, but this entails documenting
everything related to proofs, which seems like a significant endeavour
for later.

We could also re-export variable names, but I think we do not need to.

p.s. I only added Maarten as a reviewer since this primarily concerns
proofs.