feat: Enforce constraint ID and inference label are provided in propagators (#179)
To fully make proof-logging a first-class citizen of the Pumpkin solver,
we need to be able to associate every individual propagation with the
_constraint_ (not propagator) that implied the propagation. At the
moment, we assume there is a 1-to-many relationship between constraints
and propagators. I.e., every propagator propagates for exactly one
constraint. Already, there is one propagator that breaks that
assumption, namely the `NogoodPropagator`. In the future, there will
likely be more.

The solution is that every propagation and conflict requires an
`InferenceCode`, which is conceptually a tuple of a `ConstraintTag` (an
identifier for a specific constraint) and an `InferenceLabel` (an
identifier for the inference used to make the propagation).

This PR changes the interface of `PropagationContextMut` in the
following ways:
- [x] Rename `post_predicate` to `post`.
- [x] Remove `set_{lower|upper}_bound` and `remove`, and use
`post(Predicate)` everywhere. This simplifies the interface and allows
for easier refactoring in the future.
- [x] Add an `InferenceCode` parameter to `post()`. This means we
statically enforce that every propagation can be logged to a proof.

Additionally, this PR introduces the following:
- [x] We introduce the notion of `PropagatorConstructor` to replace the
`initialise_at_root` on `Propagator`.