feat: Implement reification of the table constraint
The negated version of the table constraint is a set of conflicts. We
choose to encode that as a set of nogoods, but that means there is no
arc consistency on the support variables.

MiniZinc rewrites `fzn_table_int_reif` constraints to `fzn_table_int`.
In the future we may want to explore whether we should implement the
rewrite that MiniZinc does as well, but it can potentially explode the
size of the table.

This at least allows us to support the XCSP3 'negative table constraint'
in our Python interface.