fix: fixing issue with element lazy explanation by adding trail position to explanation context (#191)
Currently, the element propagator makes use of the _current_ context
when performing its lazy explanations; however, this means that the lazy
explanation could contain elements which are **later** on the trail.

This PR addresses this issue by adding the trail position at which the
propagation which is being explained took place. This change allows us
to retrieve the domain at the time of propagation in lazy explanations.