refactor: Change the nogood inference label to print 'nogood' (#208)
Previously it printed 'nogood_label', which is the snake_case version of
the label type name. In this case it makes sense that the type name and
the actual string differ, since the name 'Nogood' for the inference
label would be confusing, whereas with 'NogoodLabel' it is clear what
the type represents.