fix: Disable recursive minimisation when logging the full proof (#215)
When performing recursive minimisation, our logging of inferences is
completely wrong. We should only log the inferences that are actually
used when removing a predicate. For now it is easier to disable it, then
we can revisit it later. Issue #214 is the reminder that we do not
forget to implement this at some point.