fix: Log the correct proofs when optimising (#216)
This PR updates Linear SAT-UNSAT optimisation to be done through
assumptions instead of a hard constraint. This assures that the proof is
also valid when optimising.

When using a hard constraint, the propagator on the objective will
generate inferences that we cannot trace to the model. Additionally,
that constraint is always changing, becoming tighter and tighter. That
issue does not exist when solving under assumptions. It very naturally
will end the proof with the statement `[obj >= val] -> false`.